+++ src/files/index.ts
@@ -1,7 +1,7 @@
-export { getFilenameByPath } from './get-filename-by-path';
-export { getFilenameExtByName } from './get-filename-ext-by-name';
-export { getFilenameWOutExtByName } from './get-filename-wout-ext-by-name';
-export { getFolderByPath } from './get-folder-by-path';
-export { sanitizeFilename } from './sanitize-filename';
-export { sanitizePath } from './sanitize-path';
-export { slugifyFilename } from './slugify-filename';
+export * from './get-filename-by-path';
+export * from './get-filename-ext-by-name';
+export * from './get-filename-wout-ext-by-name';
+export * from './get-folder-by-path';
+export * from './sanitize-filename';
+export * from './sanitize-path';
+export * from './slugify-filename';