+++ src/numbers/index.ts
@@ -1,6 +1,6 @@
-export { isNumeric, isNumberic } from './is-numeric';
-export { isEven } from './is-even';
-export { isOdd } from './is-odd';
-export { numberFixDecimals } from './number-fix-decimals';
-export { numberPrintDecimals } from './number-print-decimals';
-export { randomMinMax } from './random-min-max';
+export * from './is-numeric';
+export * from './is-even';
+export * from './is-odd';
+export * from './number-fix-decimals';
+export * from './number-print-decimals';
+export * from './random-min-max';