+++ src/php-serialize/index.ts
@@ -1,3 +1,3 @@
-export { phpIsSerialized } from './php-is-serialized';
-export { phpSerialize } from './php-serialize';
-export { phpUnserialize } from './php-unserialize';
+export * from './php-is-serialized';
+export * from './php-serialize';
+export * from './php-unserialize';